added missing copyrights, and fixed 2 error messages in jnilog.c

+++ src/Runtime/jni/jnilog.c
@@ -168,8 +168,8 @@ void log_printf(
    * of 8 bytes on z/OS.
    */
   snprintf(buf + strlen(buf), LOG_MAX_LEN - strlen(buf),
-      "[%016lx][%s]%s:%s:%d ", get_threadid(), log_level_name[level],
-      get_filename(file), func, line);
+      "[%016lx][%s]%s:%s:%d ", (unsigned long)get_threadid(),
+      log_level_name[level], get_filename(file), func, line);
 
   /* Output actual log data */
   va_list log_data;
@@ -208,7 +208,8 @@ static FILE *get_log_file_by_name(char *name) {
   else {
     char *tname = (char *)malloc(strlen(name) + 32);
     if (tname) {
-      snprintf(tname, strlen(name) + 32, "%s.%016lx", name, get_threadid());
+      snprintf(tname, strlen(name) + 32, "%s.%016lx", name,
+          (unsigned long)get_threadid());
       fp = fopen(tname, "w");
       free(tname);
     }
